Refactor action registrations into helper modules
