+++ packages/contents/src/actions/basicActions.ts
@@ -1,0 +1,244 @@
+import type { Registry } from '@kingdom-builder/protocol';
+import { Resource } from '../resources';
+import { Stat } from '../stats';
+import { DevelopmentId } from '../developments';
+import {
+	action,
+	compareRequirement,
+	developmentEvaluator,
+	effect,
+	populationEvaluator,
+	resourceParams,
+	statEvaluator,
+	developmentParams,
+	actionParams,
+	actionEffectGroup,
+	actionEffectGroupOption,
+	populationParams,
+} from '../config/builders';
+import {
+	ActionMethods,
+	DevelopmentMethods,
+	LandMethods,
+	ResourceMethods,
+	Types,
+	PopulationMethods,
+} from '../config/builderShared';
+import { Focus } from '../defs';
+import type { ActionDef } from '../actions';
+import { ActionCategory, ActionId, PopulationEvaluationId } from '../actions';
+
+export function registerBasicActions(registry: Registry<ActionDef>) {
+	registry.add(
+		ActionId.expand,
+		action()
+			.id(ActionId.expand)
+			.name('Expand')
+			.icon('🌱')
+			.cost(Resource.ap, 1)
+			.cost(Resource.gold, 2)
+			.effect(effect(Types.Land, LandMethods.ADD).param('count', 1).build())
+			.effect(
+				effect(Types.Resource, ResourceMethods.ADD)
+					.params(resourceParams().key(Resource.happiness).amount(1))
+					.build(),
+			)
+			.category(ActionCategory.Basic)
+			.order(1)
+			.focus(Focus.Economy)
+			.build(),
+	);
+
+	registry.add(
+		ActionId.overwork,
+		action()
+			.id(ActionId.overwork)
+			.name('Overwork')
+			.icon('🛠️')
+			.cost(Resource.ap, 1)
+			.effect(
+				effect()
+					.evaluator(developmentEvaluator().id(DevelopmentId.Farm))
+					.effect(
+						effect(Types.Resource, ResourceMethods.ADD)
+							.round('down')
+							.params(resourceParams().key(Resource.gold).amount(2))
+							.build(),
+					)
+					.effect(
+						effect(Types.Resource, ResourceMethods.REMOVE)
+							.round('up')
+							.params(resourceParams().key(Resource.happiness).amount(0.5))
+							.allowShortfall()
+							.build(),
+					)
+					.build(),
+			)
+			.category(ActionCategory.Basic)
+			.order(2)
+			.focus(Focus.Economy)
+			.build(),
+	);
+
+	registry.add(
+		ActionId.develop,
+		action()
+			.id(ActionId.develop)
+			.name('Develop')
+			.icon('🏗️')
+			.cost(Resource.ap, 1)
+			.cost(Resource.gold, 3)
+			.effect(
+				effect(Types.Development, DevelopmentMethods.ADD)
+					.params(developmentParams().id('$id').landId('$landId'))
+					.build(),
+			)
+			.category(ActionCategory.Development)
+			.order(1)
+			.focus(Focus.Economy)
+			.build(),
+	);
+
+	registry.add(
+		ActionId.tax,
+		action()
+			.id(ActionId.tax)
+			.name('Tax')
+			.icon('💰')
+			.cost(Resource.ap, 1)
+			.effect(
+				effect()
+					.evaluator(populationEvaluator().id(PopulationEvaluationId.tax))
+					.effect(
+						effect(Types.Resource, ResourceMethods.ADD)
+							.params(resourceParams().key(Resource.gold).amount(4))
+							.build(),
+					)
+					.effect(
+						effect(Types.Resource, ResourceMethods.REMOVE)
+							.round('up')
+							.params(resourceParams().key(Resource.happiness).amount(0.5))
+							.allowShortfall()
+							.build(),
+					)
+					.build(),
+			)
+			.category(ActionCategory.Basic)
+			.order(3)
+			.focus(Focus.Economy)
+			.build(),
+	);
+
+	registry.add(
+		ActionId.reallocate,
+		action()
+			.id(ActionId.reallocate)
+			.name('Reallocate')
+			.icon('🔄')
+			.cost(Resource.ap, 1)
+			.cost(Resource.gold, 5)
+			.category(ActionCategory.Basic)
+			.order(4)
+			.focus(Focus.Economy)
+			.build(),
+	);
+
+	registry.add(
+		ActionId.raise_pop,
+		action()
+			.id(ActionId.raise_pop)
+			.name('Hire')
+			.icon('👶')
+			.cost(Resource.ap, 1)
+			.cost(Resource.gold, 5)
+			.requirement(
+				compareRequirement()
+					.left(populationEvaluator())
+					.operator('lt')
+					.right(statEvaluator().key(Stat.maxPopulation))
+					.message('Free space for 👥')
+					.build(),
+			)
+			.effect(
+				effect(Types.Population, PopulationMethods.ADD)
+					.params(populationParams().role('$role'))
+					.build(),
+			)
+			.effect(
+				effect(Types.Resource, ResourceMethods.ADD)
+					.params(resourceParams().key(Resource.happiness).amount(1))
+					.build(),
+			)
+			.effect(
+				effect(Types.Action, ActionMethods.PERFORM)
+					.params(actionParams().id(ActionId.reallocate))
+					.build(),
+			)
+			.category(ActionCategory.Population)
+			.order(1)
+			.focus(Focus.Economy)
+			.build(),
+	);
+
+	const royalDecreeDevelopGroup = actionEffectGroup('royal_decree_develop')
+		.layout('compact')
+		.option(
+			actionEffectGroupOption('royal_decree_house')
+				.icon('🏠')
+				.action(ActionId.develop)
+				.paramDevelopmentId(DevelopmentId.House)
+				.paramLandId('$landId'),
+		)
+		.option(
+			actionEffectGroupOption('royal_decree_farm')
+				.icon('🌾')
+				.action(ActionId.develop)
+				.paramDevelopmentId(DevelopmentId.Farm)
+				.paramLandId('$landId'),
+		)
+		.option(
+			actionEffectGroupOption('royal_decree_outpost')
+				.icon('🏹')
+				.action(ActionId.develop)
+				.paramDevelopmentId(DevelopmentId.Outpost)
+				.paramLandId('$landId'),
+		)
+		.option(
+			actionEffectGroupOption('royal_decree_watchtower')
+				.icon('🗼')
+				.action(ActionId.develop)
+				.paramDevelopmentId(DevelopmentId.Watchtower)
+				.paramLandId('$landId'),
+		);
+
+	registry.add(
+		ActionId.royal_decree,
+		action()
+			.id(ActionId.royal_decree)
+			.name('Royal Decree')
+			.icon('📜')
+			.cost(Resource.ap, 1)
+			.cost(Resource.gold, 12)
+			.effect(
+				effect(Types.Action, ActionMethods.PERFORM)
+					.params(actionParams().id(ActionId.expand))
+					.build(),
+			)
+			.effect(
+				effect(Types.Action, ActionMethods.PERFORM)
+					.params(actionParams().id(ActionId.till).landId('$landId'))
+					.build(),
+			)
+			.effectGroup(royalDecreeDevelopGroup)
+			.effect(
+				effect(Types.Resource, ResourceMethods.REMOVE)
+					.params(resourceParams().key(Resource.happiness).amount(3))
+					.allowShortfall()
+					.build(),
+			)
+			.category(ActionCategory.Basic)
+			.order(5)
+			.focus(Focus.Economy)
+			.build(),
+	);
+}
